+++ Andare.cabal
@@ -26,6 +26,7 @@ source-repository head
 library
   exposed-modules:
       Lib
+      Parsing
   other-modules:
       Paths_Andare
   hs-source-dirs:
